+++ shortcuts/sheets/helpers.go
@@ -11,9 +11,7 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
 	neturl "net/url"
-	"slices"
 	"strings"
 
 	"github.com/larksuite/cli/errs"
@@ -541,278 +539,4 @@ func requireJSONArray(runtime flagView, name string) ([]interface{}, error) {
 		return nil, sheetsValidationForFlag(name, "--%s must be a JSON array", name)
 	}
 	return a, nil
-}
-
-// ─── style flags (shared by +cells-set-style and +cells-batch-set-style) ─
-
-// buildCellStyleFromFlags reads the 12 flat style flags and returns the
-// cell_styles map expected by set_cell_range. Skips any flag the user
-// didn't set so partial styles work.
-func buildCellStyleFromFlags(runtime flagView) map[string]interface{} {
-	style := map[string]interface{}{}
-	if v := runtime.Str("background-color"); v != "" {
-		style["background_color"] = v
-	}
-	if v := runtime.Str("font-color"); v != "" {
-		style["font_color"] = v
-	}
-	if v := runtime.Str("font-family"); v != "" {
-		style["font_family"] = v
-	}
-	if runtime.Changed("font-size") && runtime.Float64("font-size") > 0 {
-		style["font_size"] = runtime.Float64("font-size")
-	}
-	if v := runtime.Str("font-style"); v != "" {
-		style["font_style"] = v
-	}
-	if v := runtime.Str("font-weight"); v != "" {
-		style["font_weight"] = v
-	}
-	if v := runtime.Str("font-line"); v != "" {
-		style["font_line"] = v
-	}
-	if v := runtime.Str("horizontal-alignment"); v != "" {
-		style["horizontal_alignment"] = v
-	}
-	if v := runtime.Str("vertical-alignment"); v != "" {
-		style["vertical_alignment"] = v
-	}
-	if v := runtime.Str("word-wrap"); v != "" {
-		style["word_wrap"] = v
-	}
-	if v := runtime.Str("number-format"); v != "" {
-		style["number_format"] = v
-	}
-	return style
-}
-
-// cellStyleAliases maps shorthand cell_styles field names that models commonly
-// hallucinate (Excel / openpyxl / CSS conventions) onto the canonical field
-// names the backend expects. Only the unambiguous alignment shorthands are
-// aliased — they are the high-frequency miss; ambiguous guesses (e.g. "color",
-// "bg_color", "text_align") are intentionally left out so a wrong guess still
-// surfaces as an error rather than being silently reinterpreted.
-var cellStyleAliases = []struct{ alias, canonical string }{
-	{"horizontal_align", "horizontal_alignment"},
-	{"halign", "horizontal_alignment"},
-	{"vertical_align", "vertical_alignment"},
-	{"valign", "vertical_alignment"},
-	// wrap family: word_wrap is the sole wrap concept, no ambiguity. 07-20
-	// eval: wrap_text alone produced an 88-issue retry loop on --styles.
-	{"wrap_text", "word_wrap"},
-	{"text_wrap", "word_wrap"},
-}
-
-// cellStyleEnumFields sources the enum vocabulary for enum-bearing
-// cell_styles fields from the +cells-set-style flag-defs, so the payload path
-// (--styles / typed --cells) validates and canonicalizes values the same way
-// the cobra flag path does. 07-20 eval: "vertical_alignment":"center" (CSS
-// vocabulary; Lark spells it "middle") passed the CLI and burned a
-// server-side round trip ~10 times — the flag path had normalized it since
-// round 2, the payload path never did.
-func cellStyleEnumFields() map[string][]string {
-	defs, err := loadFlagDefs()
-	if err != nil {
-		return nil
-	}
-	spec, ok := defs["+cells-set-style"]
-	if !ok {
-		return nil
-	}
-	out := map[string][]string{}
-	for _, df := range spec.Flags {
-		if df.Kind != "own" || df.Type != "string" || len(df.Enum) == 0 {
-			continue
-		}
-		out[strings.ReplaceAll(df.Name, "-", "_")] = df.Enum
-	}
-	return out
-}
-
-// normalizeCellStyleAliases renames known shorthand keys in a single
-// cell_styles map to their canonical equivalents, in place, so a model that
-// writes e.g. "horizontal_align" instead of "horizontal_alignment" still
-// applies the style instead of hitting an "unsupported field" error (--styles)
-// or having the field silently dropped by the backend (typed --cells). If both
-// the shorthand and its canonical key are present it returns a validation error
-// rather than picking one. It then canonicalizes enum VALUES (casing + known
-// cross-vocabulary aliases like CSS "center" → Lark "middle"; boolean
-// word_wrap → the enum) and rejects off-enum values client-side instead of
-// letting the server fail the whole batch. path labels the map for errors.
-func normalizeCellStyleAliases(style map[string]interface{}, path string) error {
-	if len(style) == 0 {
-		return nil
-	}
-	for _, a := range cellStyleAliases {
-		v, ok := style[a.alias]
-		if !ok {
-			continue
-		}
-		if _, exists := style[a.canonical]; exists {
-			return common.ValidationErrorf("%s.%s conflicts with %s; pass only %s", path, a.alias, a.canonical, a.canonical)
-		}
-		style[a.canonical] = v
-		delete(style, a.alias)
-	}
-	// fore_color is deliberately NOT aliased: in openpyxl vocabulary fgColor
-	// is the FILL color while a plain reading suggests the font color — a
-	// silent pick could color the wrong thing. Prescribe both options.
-	if _, has := style["fore_color"]; has {
-		return common.ValidationErrorf("%s.fore_color is ambiguous — use font_color for text color or background_color for the cell fill", path)
-	}
-	// Boolean wrap habit: true unambiguously means wrap on, false means off.
-	if b, isBool := style["word_wrap"].(bool); isBool {
-		if b {
-			style["word_wrap"] = "auto-wrap"
-		} else {
-			style["word_wrap"] = "overflow"
-		}
-	}
-	for field, enum := range cellStyleEnumFields() {
-		raw, has := style[field]
-		if !has {
-			continue
-		}
-		val, isStr := raw.(string)
-		if !isStr || val == "" || slices.Contains(enum, val) {
-			continue
-		}
-		if canon := canonicalEnumValue(val, enum); canon != "" {
-			style[field] = canon
-			continue
-		}
-		msg := fmt.Sprintf("%s.%s value %q is invalid (allowed: %s)", path, field, val, strings.Join(enum, ", "))
-		if match := closestEnumValue(val, enum); match != "" {
-			msg += fmt.Sprintf("; did you mean %q?", match)
-		}
-		return common.ValidationErrorf("%s", msg)
-	}
-	return nil
-}
-
-// normalizeTypedCellsStyleAliases walks a typed --cells 2D array and applies
-// normalizeCellStyleAliases to every cell's inline cell_styles object, so the
-// alignment shorthands are accepted on +cells-set the same as on --styles.
-// It also expands the border "all" shorthand and intercepts border_styles
-// mis-nested inside cell_styles — both server-rejected shapes that eval
-// traces show surviving CLI validation and costing a full network round
-// trip. Structure is checked leniently to match the pass-through contract:
-// any element that isn't the expected shape is skipped, not rejected.
-func normalizeTypedCellsStyleAliases(cells []interface{}, path string) error {
-	for r, rowRaw := range cells {
-		row, ok := rowRaw.([]interface{})
-		if !ok {
-			continue
-		}
-		for c, cellRaw := range row {
-			cell, ok := cellRaw.(map[string]interface{})
-			if !ok {
-				continue
-			}
-			// cells[][].style is the habitual spelling of cell_styles (recurring
-			// server-side 900015206 in eval traces) — rewrite when unambiguous.
-			if styleObj, isObj := cell["style"].(map[string]interface{}); isObj {
-				if _, has := cell["cell_styles"]; has {
-					return common.ValidationErrorf("%s[%d][%d].style conflicts with cell_styles; pass only cell_styles", path, r, c)
-				}
-				cell["cell_styles"] = styleObj
-				delete(cell, "style")
-			}
-			// cells[][].type is not a cell field; the value type is whatever the
-			// JSON value is. Reject with the fix instead of a server round trip.
-			if _, has := cell["type"]; has {
-				return common.ValidationErrorf("%s[%d][%d].type is not a cell field — the value type is inferred from the JSON value; control display format via cell_styles.number_format", path, r, c)
-			}
-			if bs, ok := cell["border_styles"].(map[string]interface{}); ok {
-				expandBorderAllShorthand(bs)
-			}
-			st, ok := cell["cell_styles"].(map[string]interface{})
-			if !ok {
-				continue
-			}
-			if _, misNested := st["border_styles"]; misNested {
-				return common.ValidationErrorf(
-					"%s[%d][%d].cell_styles.border_styles is not valid — border_styles is a top-level cell field, a sibling of cell_styles; move it up one level",
-					path, r, c)
-			}
-			if err := normalizeCellStyleAliases(st, fmt.Sprintf("%s[%d][%d].cell_styles", path, r, c)); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// expandBorderAllShorthand rewrites the "all" side shorthand — habitual from
-// Excel / openpyxl vocabulary, rejected by the backend — into the four
-// explicit sides, in place. An explicitly set side wins over the shorthand.
-// Applied on both the typed --cells path and the --styles path, so batch
-// sub-ops get the same rewrite as standalone calls.
-func expandBorderAllShorthand(border map[string]interface{}) {
-	if all, ok := border["all"]; ok {
-		for _, side := range []string{"top", "bottom", "left", "right"} {
-			if _, exists := border[side]; !exists {
-				border[side] = all
-			}
-		}
-		delete(border, "all")
-	}
-	// Weight vocabulary in the style slot ("thin"/"medium"/"thick" are the
-	// habitual Excel words; the largest residual styles cluster in the 07-21
-	// rerun wrote them into border_styles.<side>.style of the FULL nested
-	// form). A thin border always means a thin solid line: move the word to
-	// weight and default style to solid. Only when weight is absent — an
-	// explicit conflicting weight keeps the enum error path.
-	for _, raw := range border {
-		side, ok := raw.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		s, _ := side["style"].(string)
-		switch strings.ToLower(s) {
-		case "thin", "medium", "thick":
-			if _, hasWeight := side["weight"]; !hasWeight {
-				side["weight"] = strings.ToLower(s)
-				side["style"] = "solid"
-			}
-		}
-	}
-}
-
-// borderStylesFromFlag parses --border-styles as a JSON object (top/bottom/
-// left/right with style sub-objects), expanding the "all" side shorthand the
-// same as the typed --cells and --styles paths so +cells-set-style /
-// +cells-batch-set-style don't ship {"all":…} for the backend to reject.
-// Returns nil when the flag is empty.
-func borderStylesFromFlag(runtime flagView) (map[string]interface{}, error) {
-	if runtime.Str("border-styles") == "" {
-		return nil, nil
-	}
-	v, err := parseJSONFlag(runtime, "border-styles")
-	if err != nil {
-		return nil, err
-	}
-	m, ok := v.(map[string]interface{})
-	if !ok {
-		return nil, sheetsValidationForFlag("border-styles", "--border-styles must be a JSON object")
-	}
-	expandBorderAllShorthand(m)
-	return m, nil
-}
-
-// requireAnyStyleFlag ensures at least one style-defining flag (style or
-// border) is set — otherwise the request would do nothing.
-func requireAnyStyleFlag(runtime flagView) error {
-	if len(buildCellStyleFromFlags(runtime)) > 0 {
-		return nil
-	}
-	if runtime.Str("border-styles") != "" {
-		return nil
-	}
-	return common.ValidationErrorf("at least one style flag is required (e.g. --background-color, --font-weight, --border-styles)").
-		WithParams(
-			sheetsInvalidParam("background-color", "required; specify at least one style flag"),
-			sheetsInvalidParam("font-weight", "required; specify at least one style flag"),
-			sheetsInvalidParam("border-styles", "required; specify at least one style flag"),
-		)
 }
