+++ shortcuts/sheets/lark_sheet_workbook.go
@@ -1460,139 +1460,6 @@ func normalizeWorkbookCreateStyleObject(in map[string]interface{}, path string) 
 	return out, nil
 }
 
-// foldBorderFamilyAliases rewrites the habitual flattened border vocabulary
-// (Excel / openpyxl conventions) into the canonical nested border_styles
-// object, in place. 07-20 eval: the border family alone accounted for the
-// largest --styles error cluster (borders / border / border_bottom /
-// border_style / border_top_color / …), each burning a full payload retry.
-// Accepted rewrites, all unambiguous:
-//
-//	borders / border   (object) → border_styles (side-keyed) or border_styles.all (attr-keyed)
-//	border_top|bottom|left|right (object) → border_styles.<side>
-//	border_style|color|weight  (scalar)  → border_styles.all.<attr>
-//	border_<side>_<style|color|weight> (scalar) → border_styles.<side>.<attr>
-//
-// A border_style value from the WEIGHT vocabulary (thin/medium/thick — the
-// habitual Excel word) sets weight and defaults style to solid: a "thin
-// border" always means a thin solid line. Conflicts with an explicitly given
-// border_styles error out instead of picking a side.
-func foldBorderFamilyAliases(in map[string]interface{}, path string) error {
-	sides := map[string]bool{"top": true, "bottom": true, "left": true, "right": true, "all": true}
-	attrs := map[string]bool{"style": true, "color": true, "weight": true}
-	borderWeights := map[string]bool{"thin": true, "medium": true, "thick": true}
-
-	ensureBorder := func() map[string]interface{} {
-		bs, ok := in["border_styles"].(map[string]interface{})
-		if !ok {
-			bs = map[string]interface{}{}
-			in["border_styles"] = bs
-		}
-		return bs
-	}
-	setSideAttr := func(side, attr string, v interface{}, from string) error {
-		bs := ensureBorder()
-		sideObj, ok := bs[side].(map[string]interface{})
-		if !ok {
-			if _, exists := bs[side]; exists {
-				return common.ValidationErrorf("%s.%s conflicts with border_styles.%s; keep one form", path, from, side)
-			}
-			sideObj = map[string]interface{}{}
-			bs[side] = sideObj
-		}
-		if _, exists := sideObj[attr]; exists {
-			return common.ValidationErrorf("%s.%s conflicts with border_styles.%s.%s; keep one form", path, from, side, attr)
-		}
-		sideObj[attr] = v
-		return nil
-	}
-	setSide := func(side string, v interface{}, from string) error {
-		obj, ok := v.(map[string]interface{})
-		if !ok {
-			return common.ValidationErrorf("%s.%s must be an object like {\"style\":\"solid\",\"color\":\"#000000\"}", path, from)
-		}
-		for attr, av := range obj {
-			if !attrs[attr] {
-				return common.ValidationErrorf("%s.%s.%s is not a border attribute (want style/weight/color)", path, from, attr)
-			}
-			if err := setSideAttr(side, attr, av, from); err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-	// border_style with a weight-vocabulary value means "thin solid line".
-	setAllScalar := func(attr string, v interface{}, from string) error {
-		if attr == "style" {
-			if s, ok := v.(string); ok && borderWeights[strings.ToLower(s)] {
-				if err := setSideAttr("all", "weight", strings.ToLower(s), from); err != nil {
-					return err
-				}
-				return setSideAttr("all", "style", "solid", from)
-			}
-		}
-		return setSideAttr("all", attr, v, from)
-	}
-
-	for _, key := range []string{"borders", "border"} {
-		v, has := in[key]
-		if !has {
-			continue
-		}
-		obj, ok := v.(map[string]interface{})
-		if !ok {
-			return common.ValidationErrorf("%s.%s must be an object — either side-keyed ({\"top\":{…},\"bottom\":{…}} / {\"all\":{…}}) or attribute-keyed ({\"style\":\"solid\",\"color\":\"#000\"} = all four sides)", path, key)
-		}
-		sideKeyed := false
-		for k := range obj {
-			if sides[k] {
-				sideKeyed = true
-				break
-			}
-		}
-		if sideKeyed {
-			for side, sv := range obj {
-				if !sides[side] {
-					return common.ValidationErrorf("%s.%s.%s is not a valid side (want top/bottom/left/right/all)", path, key, side)
-				}
-				if err := setSide(side, sv, key); err != nil {
-					return err
-				}
-			}
-		} else if err := setSide("all", v, key); err != nil {
-			return err
-		}
-		delete(in, key)
-	}
-	for _, side := range []string{"top", "bottom", "left", "right"} {
-		key := "border_" + side
-		if v, has := in[key]; has {
-			if err := setSide(side, v, key); err != nil {
-				return err
-			}
-			delete(in, key)
-		}
-		for attr := range attrs {
-			key := "border_" + side + "_" + attr
-			if v, has := in[key]; has {
-				if err := setSideAttr(side, attr, v, key); err != nil {
-					return err
-				}
-				delete(in, key)
-			}
-		}
-	}
-	for attr := range attrs {
-		key := "border_" + attr
-		if v, has := in[key]; has {
-			if err := setAllScalar(attr, v, key); err != nil {
-				return err
-			}
-			delete(in, key)
-		}
-	}
-	return nil
-}
-
 func workbookCreateCellStyleField(name string) bool {
 	switch name {
 	case "font_color", "font_family", "font_size", "font_weight", "font_style", "font_line",
